+++ ydb/library/testlib/helpers.h
@@ -1,0 +1,39 @@
+#pragma once
+
+#include <library/cpp/testing/unittest/registar.h>
+
+#define Y_UNIT_TEST_TWIN(N, OPT)                                                                                   \
+    template <bool OPT>                                                                                            \
+    struct TTestCase##N : public TCurrentTestCase {                                                                \
+        TTestCase##N() : TCurrentTestCase() {                                                                      \
+            if constexpr (OPT) { Name_ = #N "+" #OPT; } else { Name_ = #N "-" #OPT; }                              \
+        }                                                                                                          \
+        static THolder<NUnitTest::TBaseTestCase> CreateOn()  { return ::MakeHolder<TTestCase##N<true>>();  }       \
+        static THolder<NUnitTest::TBaseTestCase> CreateOff() { return ::MakeHolder<TTestCase##N<false>>(); }       \
+        void Execute_(NUnitTest::TTestContext&) override;                                                          \
+    };                                                                                                             \
+    struct TTestRegistration##N {                                                                                  \
+        TTestRegistration##N() {                                                                                   \
+            TCurrentTest::AddTest(TTestCase##N<true>::CreateOn);                                                   \
+            TCurrentTest::AddTest(TTestCase##N<false>::CreateOff);                                                 \
+        }                                                                                                          \
+    };                                                                                                             \
+    static TTestRegistration##N testRegistration##N;                                                               \
+    template <bool OPT>                                                                                            \
+    void TTestCase##N<OPT>::Execute_(NUnitTest::TTestContext& ut_context Y_DECLARE_UNUSED)
+
+#define Y_UNIT_TEST_QUAD(N, OPT1, OPT2)                                                                                              \
+    template<bool OPT1, bool OPT2> void N(NUnitTest::TTestContext&);                                                                 \
+    struct TTestRegistration##N {                                                                                                    \
+        TTestRegistration##N() {                                                                                                     \
+            TCurrentTest::AddTest(#N "-" #OPT1 "-" #OPT2, static_cast<void (*)(NUnitTest::TTestContext&)>(&N<false, false>), false); \
+            TCurrentTest::AddTest(#N "+" #OPT1 "-" #OPT2, static_cast<void (*)(NUnitTest::TTestContext&)>(&N<true, false>), false);  \
+            TCurrentTest::AddTest(#N "-" #OPT1 "+" #OPT2, static_cast<void (*)(NUnitTest::TTestContext&)>(&N<false, true>), false);  \
+            TCurrentTest::AddTest(#N "+" #OPT1 "+" #OPT2, static_cast<void (*)(NUnitTest::TTestContext&)>(&N<true, true>), false);   \
+        }                                                                                                                            \
+    };                                                                                                                               \
+    static TTestRegistration##N testRegistration##N;                                                                                 \
+    template<bool OPT1, bool OPT2>                                                                                                   \
+    void N(NUnitTest::TTestContext&)
+
+
